// angular/src/app/main/supplyChain/inventory/shared/dto/inventory-glLink-dto.ts
export class InventoryGlLinkDto {
    id: number | undefined;
    locId: number | undefined;
    locDesc: string | undefined;
    segId: string | undefined;
    segDesc: string | undefined;
    accRec: string | undefined;
    accRecDesc: string | undefined;
    accRet: string | undefined;
    accRetDesc: string | undefined;
    accAdj: string | undefined;
    accAdjDesc: string | undefined;
    accCgs: string | undefined;
    accCgsDesc: string | undefined;
    accWip: string | undefined;
    accWipDesc: string | undefined;
    glLocId:number|undefined;
    glLocDesc:string|undefined;
}